maTapFD is unsigned
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@12911 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/ConsoleImpl2.cpp
@@ -1163,8 +1163,8 @@ DECLCALLBACK(int) Console::configConstructor(PVM pVM, void *pvConsole)
                 if (SUCCEEDED(hrc))
                 {
 #ifdef VBOX_WITH_UNIXY_TAP_NETWORKING
-                    Assert (pConsole->maTapFD[ulInstance] >= 0);
-                    if (pConsole->maTapFD[ulInstance] >= 0)
+                    Assert ((int)pConsole->maTapFD[ulInstance] >= 0);
+                    if ((int)pConsole->maTapFD[ulInstance] >= 0)
                     {
                         if (fSniffer)
                         {
